Rename `TestWithMockContext` to `FreshContextTest`.
